Convert textures to optionals
