--- addons/entrenching/$PBOPREFIX$
@@ -1,1 +1,0 @@
-x\tacgt\addons\entrenching
