--- addons/entrenching/CfgVehicles.hpp
@@ -1,15 +1,0 @@
-class CfgVehicles {
-    class Bag_Base;
-    class CUP_O_RUS_Patrol_bag_BeigeDigital_Shovel: Bag_Base {
-        ace_trenches_entrenchingTool = 1;
-    };
-    class CUP_O_RUS_Patrol_bag_Desert_Shovel: Bag_Base {
-        ace_trenches_entrenchingTool = 1;
-    };
-    class CUP_O_RUS_Patrol_bag_Summer_Shovel: Bag_Base {
-        ace_trenches_entrenchingTool = 1;
-    };
-    class CUP_O_RUS_Patrol_bag_Green_Shovel: Bag_Base {
-        ace_trenches_entrenchingTool = 1;
-    };
-};
