--- addons/entrenching/README.md
@@ -1,3 +1,0 @@
-# Entrenching
-
-Adds ACE entrenching abilities to select backpacks.
