--- addons/entrenching/config.cpp
@@ -1,18 +1,0 @@
-#include "script_component.hpp"
-
-class CfgPatches {
-    class ADDON {
-        name = COMPONENT_NAME;
-        units[] = {};
-        weapons[] = {};
-        magazines[] = {};
-        requiredVersion = REQUIRED_VERSION;
-        requiredAddons[] = {"tacgt_main", "CUP_Weapons_Backpacks"};
-        author = ECSTRING(main,Authors);
-        authors[] = {"Mike"};
-        url = ECSTRING(main,URL);
-        VERSION_CONFIG;
-    };
-};
-
-#include "CfgVehicles.hpp"
