--- addons/entrenching/script_component.hpp
@@ -1,4 +1,0 @@
-#define COMPONENT entrenching
-#define COMPONENT_BEAUTIFIED Entrenching
-#include "\x\tacgt\addons\main\script_mod.hpp"
-#include "\x\tacgt\addons\main\script_macros.hpp"
