+++ addons/sensitivity/CfgVehicles.hpp
@@ -2,7 +2,6 @@
 
 class CfgVehicles {
     // Base Classes
-    class Civilian_F;
     class Land;
     class SoldierWB;
     class SoldierEB;
@@ -13,23 +12,15 @@ class CfgVehicles {
     };
 
     // TAC - Units
-    #include "units\tacu_units.hpp"
+    class TACU_Main_U_BLUFOR_Soldier_Base: SoldierWB {
+        MACRO_SENSITIVITY
+    };
 
-    // CUP Units
-    #include "units\cup\acr.hpp"
-    #include "units\cup\baf.hpp"
-    #include "units\cup\cdf.hpp"
-    #include "units\cup\chedaki.hpp"
-    #include "units\cup\civilians.hpp"
-    #include "units\cup\germany.hpp"
-    #include "units\cup\hil.hpp"
-    #include "units\cup\napa.hpp"
-    #include "units\cup\pmc.hpp"
-    #include "units\cup\racs.hpp"
-    #include "units\cup\russia.hpp"
-    #include "units\cup\sla.hpp"
-    #include "units\cup\taki.hpp"
-    #include "units\cup\takiinsurgents.hpp"
-    #include "units\cup\usarmy.hpp"
-    #include "units\cup\usmc.hpp"
+    class TACU_Main_U_OPFOR_Soldier_Base: SoldierEB {
+        MACRO_SENSITIVITY
+    };
+
+    class TACU_Main_U_INDEP_Soldier_Base: SoldierGB {
+        MACRO_SENSITIVITY
+    };
 };
