--- addons/sensitivity/units/cup/acr.hpp
@@ -1,3 +1,0 @@
-class CUP_Creatures_Military_ACR_Soldier_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
