--- addons/sensitivity/units/cup/baf.hpp
@@ -1,11 +1,0 @@
-class CUP_BAF_Soldier_DPM_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
-
-class CUP_BAF_Soldier_DDPM_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
-
-class CUP_BAF_Soldier_MTP_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
