--- addons/sensitivity/units/cup/cdf.hpp
@@ -1,3 +1,0 @@
-class CUP_Creatures_Military_CDF_Soldier_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
