--- addons/sensitivity/units/cup/chedaki.hpp
@@ -1,3 +1,0 @@
-class CUP_Creatures_Military_OPFINS_Soldier_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
