--- addons/sensitivity/units/cup/civilians.hpp
@@ -1,8 +1,0 @@
-// cup_civil_russia inherits from these units.
-class CUP_Creatures_Civil_Chernarus_Base: Civilian_F {
-    MACRO_SENSITIVITY
-};
-
-class CUP_Creatures_Civil_Takistan_Base: Civilian_F {
-    MACRO_SENSITIVITY
-};
