--- addons/sensitivity/units/cup/germany.hpp
@@ -1,3 +1,0 @@
-class CUP_Creatures_Military_GER_Soldier_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
