--- addons/sensitivity/units/cup/hil.hpp
@@ -1,15 +1,0 @@
-class CUP_Creatures_Military_HIL_Man_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
-
-class CUP_Creatures_Military_HIL_Recon_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
-
-class CUP_Creatures_Military_HIL_Reservist_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
-
-class CUP_Creatures_Military_HIL_SF_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
