--- addons/sensitivity/units/cup/napa.hpp
@@ -1,3 +1,0 @@
-class CUP_Creatures_Military_NAPA_Soldier_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
