--- addons/sensitivity/units/cup/pmc.hpp
@@ -1,7 +1,0 @@
-class CUP_Creatures_Military_PMC_Soldier_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
-
-class CUP_Creatures_Military_PMC_Soldier_Winter_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
