--- addons/sensitivity/units/cup/racs.hpp
@@ -1,3 +1,0 @@
-class CUP_Creatures_Military_RACS_Soldier_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
