--- addons/sensitivity/units/cup/russia.hpp
@@ -1,3 +1,0 @@
-class CUP_Creatures_Military_RUS_Soldier_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
