--- addons/sensitivity/units/cup/sla.hpp
@@ -1,3 +1,0 @@
-class CUP_Creatures_Military_SLA_Soldier_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
