--- addons/sensitivity/units/cup/taki.hpp
@@ -1,3 +1,0 @@
-class CUP_Creatures_Military_TK_Soldier_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
