--- addons/sensitivity/units/cup/takiinsurgents.hpp
@@ -1,7 +1,0 @@
-class CUP_Creatures_Military_TKG_Guerrilla_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
-
-class CUP_Creatures_Military_TKI_Insurgent_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
