--- addons/sensitivity/units/cup/usarmy.hpp
@@ -1,3 +1,0 @@
-class CUP_Creatures_Military_USA_Soldier_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
