--- addons/sensitivity/units/cup/usmc.hpp
@@ -1,7 +1,0 @@
-class CUP_Creatures_Military_USMC_Soldier_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
-
-class CUP_Creatures_Military_FR_Soldier_Base: SoldierWB {
-    MACRO_SENSITIVITY
-};
