+++ addons/special/CfgWeapons.hpp
@@ -5,8 +5,6 @@ class CfgWeapons {
     class Uniform_Base;
     class UniformItem;
     class U_I_C_Soldier_Bandit_2_F;
-    class HeadgearItem;
-    class ItemCore;
     class H_HelmetB_TI_arid_F;
 
     // Ethan
@@ -15,33 +13,6 @@ class CfgWeapons {
         scope = 2;
         hiddenSelectionsTextures[] = {
             QPATHTOF(data\h_helmetb_ti_ethan_f_co.paa)
-        };
-    };
-
-    // Jack
-    class CUP_H_C_Beret_01: ItemCore {
-        class ItemInfo: HeadgearItem {
-            class HitpointsProtectionInfo {
-                class Head;
-            };
-        };
-    };
-
-    class CLASS(Jack_Beret): CUP_H_C_Beret_01 {
-        author = "GilleeDoo";
-        scope = 2;
-        displayName = "Beret, True Freedom Edition";
-        descriptionShort = "Fitted with a steel plate";
-        hiddenSelectionsTextures[] = {
-            QPATHTOF(data\jack_beret_co.paa)
-        };
-        // Mythical head protection of an unknown origin..
-        class ItemInfo: ItemInfo {
-            class HitpointsProtectionInfo: HitpointsProtectionInfo {
-                class Head: Head {
-                    armor = 10;
-                };
-            };
         };
     };
