+++ optionals/headgear/CfgWeapons.hpp
@@ -27,4 +27,33 @@ class CfgWeapons {
             QPATHTOF(data\cap_def_nymets_co.paa)
         };
     };
+
+    // Jack
+    class ItemCore;
+    class HeadgearItem;
+    class CUP_H_C_Beret_01: ItemCore {
+        class ItemInfo: HeadgearItem {
+            class HitpointsProtectionInfo {
+                class Head;
+            };
+        };
+    };
+
+    class CLASS(Jack_Beret): CUP_H_C_Beret_01 {
+        author = "GilleeDoo";
+        scope = 2;
+        displayName = "Beret, True Freedom Edition";
+        descriptionShort = "Fitted with a steel plate";
+        hiddenSelectionsTextures[] = {
+            QPATHTOF(data\jack_beret_co.paa)
+        };
+        // Mythical head protection of an unknown origin..
+        class ItemInfo: ItemInfo {
+            class HitpointsProtectionInfo: HitpointsProtectionInfo {
+                class Head: Head {
+                    armor = 10;
+                };
+            };
+        };
+    };
 };
