+++ optionals/headgear/README.md
@@ -1,3 +1,5 @@
 # Headgear
 
+Now optional, no longer using CUP.
+
 Adds an OpsCore FAST Helmet for medics and caps.
