+++ optionals/headgear/config.cpp
@@ -6,10 +6,15 @@ class CfgPatches {
         units[] = {};
         weapons[] = {
             QCLASS(OpsCore_Medic),
-            QCLASS(Cap_DEF_NYMets)
+            QCLASS(Cap_DEF_NYMets),
+            QCLASS(Jack_Beret)
         };
         requiredVersion = REQUIRED_VERSION;
-        requiredAddons[] = {"tacgt_main", "CUP_Creatures_Extra_OpsCore"};
+        requiredAddons[] = {
+            "tacgt_main",
+            "CUP_Creatures_Extra_OpsCore",
+            "CUP_Creatures_People_Civil_Chernarus",
+        };
         author = ECSTRING(main,Authors);
         authors[] = {"GilleeDoo, Mike"};
         url = ECSTRING(main,URL);
