+++ optionals/headgear/data/cup_opscore_medic_1_co.paa
