+++ optionals/headgear/data/cup_opscore_medic_2_co.paa
