+++ optionals/headgear/data/cup_opscore_medic_co.paa
