+++ optionals/headgear/script_component.hpp
